delete cli and change structure
