+++ crawler/src/main/java/ir/jimbo/crawler/App.java
@@ -4,6 +4,7 @@ package ir.jimbo.crawler;
 import ir.jimbo.crawler.config.AppConfiguration;
 import ir.jimbo.crawler.config.KafkaConfiguration;
 import ir.jimbo.crawler.config.RedisConfiguration;
+import ir.jimbo.crawler.parse.AddPageToKafka;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -46,9 +47,13 @@ public class App {
             return;
         }
 
+        parserThreads = new Thread[100];
+
+        for (int i = 0; i < 100; i++) {
+            parserThreads[i] = new Thread(new AddPageToKafka());
+            parserThreads[i].start();
+        }
+
         CacheService cacheService = new CacheService(redisConfiguration);
-        Parser parser = new Parser(appConfiguration, cacheService, kafkaConfiguration);
-        new Cli(kafkaConfiguration).start();
-        new LinkConsumer(kafkaConfiguration).startGetLinks(cacheService);
     }
 }
